{Feature} Sophus pybind integration
Summary:
add sophus pybind for SO3 and the unit test
SE3 is added to avoid CI errors , will be updated in the next diff.

Overview
SO3Group is a class of a vector of SO3 objects. We create python binding to SO3Group so that vectorized operation can be realized naturally.
This interface design is similar to scipy.Spatial.Rotation.
* slicing will be added in later diffs.
* broadcasting of N to N rotations is not added in this diff

Reviewed By: Dvad

Differential Revision: D47855520

fbshipit-source-id: 126e4c7b1b763fdc4e59bf18c2d9c30a9e648b10