{Feature} Move filenames and file keys to separate header files
Summary:
Currently all file names and keys are hardcoded into the cpp files.
This was done because it was only expected to be used within those cpp files.
However, when trying to build a file exporter, we want to be able to reuse the same definitions of the filenames.
By moving these definitions to a header file, we can simply include those during the data generation.
Exposing these also allow others to write their own loaders without having to manually copy the definitions over (which may break if we change any definitions)

Reviewed By: SeaOtocinclus

Differential Revision: D49254020

fbshipit-source-id: 95a8962c70b7047a82e400c54e956f2643db6ccc