{Feature} ADT | Add notebook tutorial for generating pointclouds from depth maps data
Summary: In this tutorial, we will explain how to generate 3D pointclouds from the ADT depth maps. This includes unprojecting points using the camera model to get the 3D ray associated with each pixel, using the depth map to compute each point's 3D coordinates in the camera frame, using the Aria pose and calibration to compute the point coordinates in the Scene frame, and finally using the RGB images to colorize the pointcloud.

Reviewed By: PiotrBrzyski, SeaOtocinclus

Differential Revision: D51130476

fbshipit-source-id: 0cde16dea958902e770f9956af0126a4ba5c568d