{Bugfix} Add warning for corrupted sequences
Summary:
We recently discovered 3 datasets that have corrupted IMU data. The remainder of the data in these VRS file are okay, but this could cause issues with users so we want to flag it.
Note that a later fix will come that will remove these datasets from the release, but we wanted to land a quick intermediate fix for now.
The current fix just outputs a warning if any of these 3 datasets are detected.
Result:

```
loading vrs and ground truth data...
[ProgressLogger][INFO]: 23153.747: Opening diskfile file 2023-11-28 17:12:01...
[VrsDataProvider][INFO]: streamId 211-1/camera-et activated
[VrsDataProvider][INFO]: streamId 214-1/camera-rgb activated
[VrsDataProvider][INFO]: streamId 247-1/baro0 activated
[VrsDataProvider][INFO]: Timecode stream found: 285-2
[VrsDataProvider][INFO]: streamId 1201-1/camera-slam-left activated
[VrsDataProvider][INFO]: streamId 1201-2/camera-slam-right activated
[VrsDataProvider][INFO]: streamId 1202-1/imu-right activated
[VrsDataProvider][INFO]: streamId 1202-2/imu-left activated
[VrsDataProvider][INFO]: streamId 1203-1/mag0 activated

[WARNING] BAD DATASET DETECTED
[AriaDigitalTwinDataProvider][ERROR]: Dataset Apartment_release_clean_seq139 has been flagged as corrupted, please use data carefully. Reason: IMU data corrupted.
[AriaDigitalTwinDataProvider][INFO]: loading instance info from json file /home/nickcharron/data_local/Apartment_release_clean_seq139/1WM103600M1292_optitrack_release_clean_seq139/instances.json
Loaded #closed loop trajectory poses records: 2897
```

Reviewed By: PiotrBrzyski

Differential Revision: D51644112

fbshipit-source-id: 809a1686b76c74aa6c121546dc7b79c34028cdb1