{Feature} MPS - Add EyeGazeVergence support
Summary: This diff adds support for reading eye gaze vergence data from an MPS file. It includes changes to the `EyeGazeReader` class to read the new data, as well as changes to the `EyeGaze` and `EyeGazeVergence` structs to represent the new data. It also includes a new test case to verify that the reader works correctly.

Differential Revision: D54089114

fbshipit-source-id: 890764c0841ccafefa32707a37f573a6e00b0ceb