{BugFix} Remove internal module definition for StreamId and RecordableTypeId
Summary:
The ADT data provider has a function to return all aria streams which has pybindings.
Unfortunately this pybinding does not have tests to ensure this works, and we recently learned that D50827432 broke it.

It turns out projectaria_tools core module defines StreamId and RecordableTypeId as local modules to avoid conflict with pyvrs (see D50827432).
This means that any python modules that import core cannot use StreamId or RecordableTypeId which is what ADT was trying to do. This could affect any tooling in the future that may depend on projectaria_tools.core

We did some searching and it doesn't look like StreamId or RecordableTypeId are defined in the current pyvrs. If this passes CI, then it should be good to land.

Reviewed By: YLouWashU

Differential Revision: D57060097

fbshipit-source-id: 3e8b2c18e8cfc40377a80368eccb1f7bdcb15ce3