{BugFix} Remove get_aria_all_streams pybinds
Summary:
The ADT data provider has a function to return all aria streams which has pybindings.
Unfortunately this pybinding was never tested, and we recently learned that it does not work.

It turns out projectaria_tools core module defines StreamId and RecordableTypeId as local modules to avoid conflict with pyvrs (see D50827432).
This means that any python modules that import core cannot use StreamId or RecordableTypeId which is what ADT was trying to do.
The temporary solution is to remove that function from ADT. This shouldn't cause any breaking changes as this was never working to begin with.
The long term solution is to not re-define the bindings for these two classes, instead to depend on pyvrs. This is left as a todo

Reviewed By: PiotrBrzyski

Differential Revision: D56856754

fbshipit-source-id: 882de1d9882c7ac3b050f651daea53365f357c09