{Build} Continuous Integration - cibuildwheel version
Summary:
Fix cibuildwheel version to 2.17.0

Last update if making the workflow using 2.18.0 and it is breaking the cibuildrepairwheel steps.

For an unknown reason, cibuildwheel 2.18.0 is breaking some assumption about dylib mac compatibiltity for MacOS. We will stick to 2.17.0 for the time being, and replace all brew install with on-demand library build in the future.

Reviewed By: vijay609

Differential Revision: D57287270

fbshipit-source-id: f3301f8bd055cbdad85ab4ef26dad45d2ce5b819