{Feature} Projects - ADT - Add bounds check whenever time variant data is queried
Summary:
There is a recurring issue where users query for data using image timestamps outside the range of available ground truth data.
See: https://github.com/facebookresearch/projectaria_tools/issues/107

Background: All ADT sequences start collecting Aria data prior to starting the tracking system.
This is because it takes a few seconds for Aria to get up and running and for the operators to disconnect Aria and get into view of the Optitrack cameras.
When you query for any data using the ADT data provider, the default is to look for the nearest result and therefore you might get data with large time differences if the query timestamp is before or after the GT data.
The solution to this is to always check that your query timestamp is within the range of the ground truth data using these two functions: get_start_time_ns, get_end_time_ns.

We show this in the tutorials, but it still seems to be an issue that many people encounter.
So to save time trying to explain this to people, this diff adds a check in the data provider which checks the bounds each time data is queried, and returns a warning if it's outside the bounds

Reviewed By: PiotrBrzyski

Differential Revision: D58438355

fbshipit-source-id: 7774d6d672fcf3fcc063f619ddbc400a9816a852