{Feature} VrsDataProvider - Resolve colab issue: imports out of order
Summary: The notebook passes in an environment where projectaria_tools is already installed. However, on collab, projectaria_tools is pip installed as a side-effect of downloading the sample files (this is common across all the sample notebooks). Therefore, the import of projectaria_tools had to move after the file-download section to work on colab.

Reviewed By: xiaqingp

Differential Revision:
D58539746

Privacy Context Container: L1262222

fbshipit-source-id: 51be768e248754835532c5ea1d45e67a14fc10b9