{Feature} Projects - Release new general dataset downloader
Summary:
With the effort to standardize project Aria datasets, we will have standard data hosting setups that allow us to standardize the download tools as well.
This diff adds a general dataset downloader meant to work with any Aria dataset. For now, we only onboard ADT but plan on onboarding other datasets to use this general tool as well.

Reviewed By: xiaqingp

Differential Revision: D58144320

fbshipit-source-id: ea0d7674c963a1a7a950fac1ab7940df7971dc41