{Deprecation} Deprecating old ADT downloader and challenge data
Summary:
We are switching to a new standard dataset downloader tool: aria_dataset_downloader
This diff deprecates the ADT specific downloader and tells users to use the new one.

This diff also starts the deprecation process for ADT challenge data.
We are leaving the challenge data downloader active, but users will need to request the cdn file to get access to the data.
Information to request the cdn file for challenge data will be added to the wiki

We also remove the dataset metadata searcher which is no longer needed since it's replace by the dataverse explorer

Reviewed By: PiotrBrzyski

Differential Revision: D58309218

fbshipit-source-id: 475e12218d7d28ff21668f3fa30ef723488d0ee5