{Visualization} ticsync_tutorial post-CVPR - Rerun in notebook
Summary:
Adds rerun in-notebook to the shipped ticsync_tutorial.ipynb

Add `__repr__` print to the metadata object for more natural display in the notebook (attempted to add `__repr__` and `__str__` to `enum MetadataTimeSyncMode`, but I could not get it to trigger in the notebook. Perhaps it's due to some limitation in PyBind. I'll investigate, but it did not affect the narrative and did not seem to be worth delaying the update.

Add type annotations to functions and function parameteers (arguments), but not to all local variables becase such seemed less than helpful when I tried it

Fixed several small nits.

Addressed clang-tidy issue.

Reviewed By: PiotrBrzyski

Differential Revision: D58841149

fbshipit-source-id: 71c7f54453ba03d43a6bc6bf5e586b068f1b90ce