Fix the handling of the different image sizes for online camera calibration
Summary:
The old mps online calibration file don't expose the image size as part of it.

But the online calirbation when it is enabled is based on the vrs image size and projection models are alraedy scaled.

Read the image size properly to make the camera calibration read from onlince calibration valid.

Reviewed By: vijay609

Differential Revision: D63001871

fbshipit-source-id: d3df61faf7acd2897ee4ec608c68261fa4e442ae