{Feature} Python - Sample - Add RGB visibility of semidense observations
Summary:
Refactor the code samples to add:
- RGB visibility and tracks of the semi dense observations

As RGB point visibility are not pre-computed, we are estimating them by computing the points that are visible at a given timestamp from the points seen by the two slam cameras.

We are introducing here a code refactoring allowing to reuse the code structure  use to compute point visibilities for 3rd party applications.
- PointsAndObservationsManager (a class to ease retrieval and manipulation of the visible point in each camera stream)
- TracksManager (a class to update and clean tracks related to MPS Point observations)

Reviewed By: maieneuro

Differential Revision: D65151407

fbshipit-source-id: 4b9c1c3a35f101bf1f176f00c6e75dfbca6acfbc