{Documentation} Clarify Python version requirements
Summary: 3.8 won't work anymore, and 3.12 is too much. I went with "A version of Python somewhere from 3.9 to 3.11", which sounds casual but avoids the traps other wordings might fall into. For instance, saying "Python 3.9 to 3.11" could be interpreted as the user needing access to that entire range. Saying "A version between 3.9 and 3.11" could be interpreted as non-inclusive, thus limiting the user to 3.10. Using mathematical symbols (">=", etc.) would be awkward visually but is still a viable alternative if we decide we don't like this wording.

Reviewed By: SeaOtocinclus

Differential Revision: D67358429

fbshipit-source-id: 0b1be96507976aeb8dab8009a7387b36764f8b7c