{Feature} Image - copy to pixel frame util (#208)
Summary:
Pull Request resolved: https://github.com/facebookresearch/projectaria_tools/pull/208

The image data within ImageDataAndRecord need to be updated.  The image data is managed by PixelFrame. The outputs from the Devignetting and Color Correction APIs are in the form of ManagedImageVariant. To address this, an image utility was created.

It is important to note that PixelFrame contains additional metadata that cannot be derived solely from ManagedImageVariant. Therefore, only the underlying image data is replaced while preserving the metadata.

Reviewed By: chpeng-fb

Differential Revision: D70644939