{Feature} C++ - Core - Make VrsDataProvider friend class of SensorData (#203)
Summary:
Pull Request resolved: https://github.com/facebookresearch/projectaria_tools/pull/203

Make `VrsDataProvider` a friend class of `SensorData` and change `dataVariant_` to a protected variable. The private variables are split into two parts because `SensorDataVariant` needs to be positioned between `StreamId` and `SensorDataType` to maintain the order of the existing initialization list. This ensures no changes to the API are introduced.

Reviewed By: hxtang

Differential Revision: D70929206

fbshipit-source-id: 52bfb1ebd8d2b327c84ad542d702eed71fa3556f