{feature} Devignetting - 3 channels masks - Devignetting (#206)
Summary:
Pull Request resolved: https://github.com/facebookresearch/projectaria_tools/pull/206

##Devignetting (Devignetting.h, Devignetting.cpp)
This update modifies the input for the devignetting function, replacing Eigen::MatrixXf with ManagedImage3F32.

To avoid differentiate the 1 & 3 channels mask, both RGB and SLAM masks binary saved three channel infor. But every channel for SLAM image is the same.

## DeviceCalibration
This update modifies the input for the loadDevignettingMask function in DeviceCalibration, replacing Eigen::MatrixXf with ManagedImage3F32.
To avoid differentiate the 1 & 3 channels mask, both RGB and SLAM masks binary saved three channel info. But every channel for SLAM image is the same.

## VrsDataProvider
This update modifies the devignetting mask used in VrsDataProvider, replacing Eigen::MatrixXf with ManagedImage3F32.
Fixed bugs for streamId.getTypeName() to correct type name.

## notebook and doc
The binary download link for vignetting mask will be updated the following diff. Therefore, the doc and notebook need to be updated at the same time.

Reviewed By: chpeng-fb

Differential Revision: D71150778

fbshipit-source-id: a756143ed35a2c11378f72a64b12ce4864c1aed0