{Feature} Core - Expose HandTrackingResult as Python API
Summary:
Expose `HandTrackingResult` as Python API. This is the Python binding regarding the newly introduced struct representing the hand tracking results. This struct is to replace the `WristAndPalmPose` that is to be deprecated in the future.

We expose the following fields in this new struct
* Hand confidence. This is the same as what's already available in WristAndPalm struct.
* Hand landmarks (`landmark_positions_device`) with their meaning specified in enum class `HandLamark` and the connectivity among them in `kHandJointConnections`. A total of 21 landmarks are provided.
* Wrist transform (`T_device_wrist`). The rotation and position of the local reference frame of hand at wrist joint. This could be useful for users to define points and compute other quantities in local hand (wrist) space, and convert to device space.
* Wrist and palm normal (`wrist_and_palm_normal_device`). This is the same as what's already available in WristAndPalm struct.

Wrist and palm information can be deduced by the new landmark positions indexed by `HandLamark.WRIST` and `HandLamark.PALM_CENTER`. We also provide two helper functions `get_wrist_position_device` and `get_palm_position_device` to make transition to this new struct easier.

Reviewed By: YLouWashU

Differential Revision: D73168721

fbshipit-source-id: df8d144b3be7aa02a0a3eb059bf1e9fabef086c0