{BugFix} [Core] - Set per-hand hand-tracking result optional if hand confidence is below zero
Summary:
CSV files record default values for hand tracking results (-1 for confidence, 0s for positions, and identify for rotation) when hand is not found. The reader will read them and create additional storage to hold the data, but they are never supposed to be used as confidence is -1. This code fixes the reader.

Users should be aware that if their original code didn't use the data structure as it should be used by checking the availability of the per-hand result and then access the values, a crash might be spotted after updating to this commit. The proper way to access the per-hand data is

C++
```
handTrackingResult = ... // Get hand tracking result for a particular frame
if (handTrackingResult.leftHand.has_value() {
  const auto& leftHandTrackingResult = handTrackingResult.leftHand.value();
  // Do something with leftHandTrackingResult
}
if (handTrackingResult.rightHand.has_value() {
  const auto& rightHandTrackingResult = handTrackingResult.rightHand.value();
  // Do something with rightHandTrackingResult
}
```

Python
```
hand_tracking_result = ... # Get hand tracking result for a particular frame
if hand_tracking_result.left_hand is not None:
  # Do something with hand_tracking_result.left_hand
if hand_tracking_result.right_hand is not None:
  # Do something with hand_tracking_result.right_hand
```

Reviewed By: danielyan86129

Differential Revision: D74088097

fbshipit-source-id: be31a28e9275b643b78a33c8d013de514c52f82a