{Feature} Core - Introduce a new hand tracking result struct for hand landmarks, wrist transform and wrist and palm normals
Summary:
Introduce a new structure `HandTrackingResult` as the future container for hand tracking outcome. This class will replace `WristAndPalmPose` in the future, and is recommended to be used for hand tracking result.

We expose the following fields in this new struct
* Hand confidence. This is the same as what's already available in WristAndPalm struct.
* Hand landmarks (`Landmarks`) with their meaning specified in enum class `HandLamark` and the connectivity among them in `kHandJointConnections`. A total of 21 landmarks are provided.
* Wrist transform (`T_Device_Wrist`). The rotation and position of the local reference frame of hand at wrist joint. This could be useful for users to define points and compute other quantities in local hand (wrist) space, and convert to device space.
* Wrist and palm normal. This is the same as what's already available in WristAndPalm struct.

Wrist and palm information can be deduced by the new landmark positions indexed by `HandLamark::WRIST` and `HandLamark::PALM_CENTER`. We also provide two helper functions `getWristPositionInDevice` and `getPalmPositionInDevice` to make transition to this new struct easier.

Reviewed By: YLouWashU

Differential Revision: D73066209

fbshipit-source-id: 297c7f5fde150d6a27eca5bc6b23b131154925fb