{Feature} Visualization - Update hand tracking visualization to better show left and right hands and different landmarks
Summary: With the addition of new landmarks. It's better to use multiple colors for left and right hands so it's easy to see which landmarks belong to which hand. esp. when fingers are crossing each other.  We also show wrist and palm in different color so it's easy to notice and backward compatible with existing visualization.

Reviewed By: YLouWashU

Differential Revision: D74381094

fbshipit-source-id: beff69807dc55acd17f497b0c04a099a9af87283