Clear up missing AEA step
Summary: Formerly, there was a "Step 4" under "Running the Jupyter Notebook locally". Then, that "Step 4" was deleted, leaving a blank between steps 3 and 5. After inquiring about this with Pierre, I changed the current step 5 to step 4. The only other thing to be noted is the "Step 1" paragraph. You'll notice it refers to skipping right ahead to Step 4 in certain situations. Now, the missing Step 4 is the one about downloading the sample sequence, which is no longer a thing. However, skipping instead to Step 3 doesn't make sense, as it would mean installing projectaria-tools, and the stipulation for skipping is that projectaria-tools has already been installed. Thus, I'm going to assume that the text remains accurate and that skipping ahead to the current Step 4 (running the tutorial) is indeed fair if projectaria-tools has already been installed.

Reviewed By: SeaOtocinclus

Differential Revision: D75905926

fbshipit-source-id: 8e3f17b49e8e4dd17cc0037bc518722ae0684f6f