{BugFix} mps_cli - Authentication - Handle Exception in case of invalid auth token
Summary:
This change fix the issue when the user install aria studio but the token is invalid or expired in the saved cache.
This diff is a bug fix for the authentication process in the Aria MPS CLI. The code changes in the authentication.py file handle exceptions in case of invalid auth tokens. The changes include catching exceptions such as ValueError, KeyError, OSError, and any other unexpected exceptions. If an exception occurs, the auth token cache file is removed and the function returns False. This ensures that the authentication process is not affected and the user is prompted to login again.

Reviewed By: maieneuro

Differential Revision: D78013372

fbshipit-source-id: c20fb91edc9f21fb4179c86761af1ea83782f25d