{BugFix} mps_cli: changing feedback id key in http helper to match the correct key (user_feedback) in the query
Summary: This diff contains a bug fix in the `constants.py` file and a code change in the `http_helper.py` file. The bug fix involves adding a new key `KEY_USER_FEEDBACK` to the `constants.py` file. The code change in `http_helper.py` involves adding the new key `KEY_USER_FEEDBACK` to the `input` dictionary when making a request to the MPS API. The changes are necessary to ensure that the `feedback_id` is correctly passed to the MPS API.

Reviewed By: PiotrBrzyski

Differential Revision: D83054863

fbshipit-source-id: 8a15e68e825f862eef228c958fc153a87f236983