{Feature} mps_cli - Recognize device type for single recording processing
Summary:
## Summary

This diff is the first of a two-part series that introduces the recognition of the recording version for single recording processing in the ARIA MPS CLI. The changes in this diff lay the groundwork for future development.

### Key Changes

*   Added a new constant `KEY_DEVICE_TYPE` to `constants.py` to represent the device type in the MPS request.
*   Introduced a new enum `MpsAriaDevice` in `types.py` to define the device types supported by the MPS.
*   Updated the `http_helper.py` file to include the `KEY_DEVICE_TYPE` constant in the `KEYS_TO_INCLUDE` set.
*   Updated the `single_recording_mps.py` file to determine the device type recording was taken with and set the `MpsAriaDevice` enum to represent it.

### Impact

These changes enable the MPS CLI to recognize and process recordings from different devices, paving the way for future development and expansion of the MPS ecosystem.

Reviewed By: vijay609

Differential Revision: D82633112

fbshipit-source-id: a3aa34efeefb409fb03ef8a4465f1312617a5de8