{Feature} update decoding to release GIL to allow multi-threading in c++
Summary:
Context
Python uses global interpretor lock (GIL) so that multi-threading in c++ becomes a single thread by locking other threads and only allow one thread to run at any given time.
This forces the decoding pybind to become serialized and can become a bottle neck.

Solution
Release the GIL only during decoding part. when invoking this code, the GIL is released so that the threads are not blocked.

Reviewed By: kongchen1992

Differential Revision: D83271679