{BugFix} crash fix: adapt magnetometerCalibrationToJsonto behave like parseMagnetometerCalibrationFromJson (#290)
Summary:
As is the load/save functions behave differently, and the result is that loading a DeviceCalibration json crashes because of magnetometer matrix positivity checks.

Assuming `parseMagnetometerCalibrationFromJson` is the ultimate source of truth (as it has Gen1/Gen2 different behaviors), this diff makes `magnetometerCalibrationToJson` do exactly the inverse while saving the json file.

Pull Request resolved: https://github.com/facebookresearch/projectaria_tools/pull/290

Reviewed By: YLouWashU

Differential Revision: D85061740

Pulled By: SeaOtocinclus

fbshipit-source-id: 350a3e2a1f0fd9a3190c585d9f38de9c603a587d