{Feature} C++ - save/load camera config to avoid read-adapting (#294)
Summary:
When loading from VRS, the factory calibration is adapted to the current resolution *ASSUMING* that those values are for a default resolution (2880 x 2880 for RGB camera).
Therefore the process of saving and reloading from JSON is currently broken, as the scaling would be applied TWICE.
The current diff makes sure the config data are saved and used if available. if not present, the default is assumed (like when loading from VRS) and loaded as usual.
This is required to make VI-BA work.

Pull Request resolved: https://github.com/facebookresearch/projectaria_tools/pull/294

Reviewed By: hxtang

Differential Revision: D85343896

Pulled By: chpeng-fb

fbshipit-source-id: 143298d3657e215c232581705a8d85c4ccbc462f