{Feature} Calib jacobians (#304)
Summary:

Add Jacobians wrt the calibration parameters.
This is required to make VI-BA work.

This diff also changes some return types to const references, as it is better practice to avoid creating temporary objects with dynamic allocation, and allows Mut fields so that the CameraCalibration type can be used in optimization.


Differential Revision: D85353427

Pulled By: maurimo