{Feature} Calib jacobians (#304)
Summary:
Pull Request resolved: https://github.com/facebookresearch/projectaria_tools/pull/304

Add Jacobians wrt the calibration parameters.
This is required to make VI-BA work.

This diff also changes some return types to const references, as it is better practice to avoid creating temporary objects with dynamic allocation, and allows Mut fields so that the CameraCalibration type can be used in optimization.

Pull Request resolved: https://github.com/facebookresearch/projectaria_tools/pull/293

Reviewed By: SeaOtocinclus

Differential Revision: D85353427

Pulled By: maurimo

fbshipit-source-id: 36b76b2d92869260d0f9ba8d60207ab2d3d9e4d1